tests/wpt/tests/css/css-properties-values-api/animation/custom-property-animation-time.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.css-houdini.org/css-properties-values-api-1">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/utils.js"></script>
<div id="target"></div>
<script>

animation_test({
  syntax: "<time>",
  inherits: false,
  initialValue: "0s"
}, {
  keyframes: ["100s", "200s"],
  expected: "150s"
}, 'Animating a custom property of type <time>');

animation_test({
  syntax: "<time>",
  inherits: false,
  initialValue: "100s"
}, {
  keyframes: "200s",
  expected: "150s"
}, 'Animating a custom property of type <time> with a single keyframe');

animation_test({
  syntax: "<time>",
  inherits: false,
  initialValue: "100s"
}, {
  composite: "add",
  keyframes: ["200s", "300s"],
  expected: "350s"
}, 'Animating a custom property of type <time> with additivity');

animation_test({
  syntax: "<time>",
  inherits: false,
  initialValue: "100s"
}, {
  composite: "add",
  keyframes: "300s",
  expected: "250s"
}, 'Animating a custom property of type <time> with a single keyframe and additivity');

animation_test({
  syntax: "<time>",
  inherits: false,
  initialValue: "100s"
}, {
  iterationComposite: "accumulate",
  keyframes: ["0s", "100s"],
  expected: "250s"
}, 'Animating a custom property of type <time> with iterationComposite');

</script>
